wlr_keyboard_group: introduce key_changed event
This introduces the key_changed event for wlr_keyboard_group. This event
is emitted when a keyboard is added to or removed from the group while
a key is currently being pressed that is not currently pressed by any
other key in the group. The purpose of this event is to allow the
compositor to update its internal state so that it does not become
corrupt. Additionally, is the key is removed from the keyboard group,
its also possible that a surface was sent the corresponding pressed
event and needs a release event sent to prevent its state from going
corrupt.

This was previously being emitted as a normal key event. However, this
is not a normal key event. There is no actual key press or release with
this event. It's purely needed for state keeping. Using a separate event
allows the compositor to handle it differently.

This event should not trigger any bindings as there is no actual key
press or release. The compositor should just update its state and notify
the surface of a key release, if needed.